+++ src/main/java/io/castle/client/model/CastleRuntimeException.java
@@ -3,8 +3,7 @@ package io.castle.client.model;
 import okhttp3.Response;
 
 /**
- * Exception thrown if the client is configured to use a THROW
- * {@code AuthenticateFailoverStrategy} and if a call to the authenticate endpoint of the Castle API fails.
+ * Runtime exception wrapping errors raised while calling the Castle API.
  */
 public class CastleRuntimeException extends RuntimeException {